♻️ Fix: 모달창 토글 커스텀 훅으로 분리
