+++ src/components/pages/QuestionPage.jsx
@@ -1,4 +1,3 @@
-import { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import useQuestionData from '../../hooks/useQuestionData';
 import FloatingButton from '../atoms/Button/FloatingButton/FloatingButton';
@@ -6,29 +5,13 @@ import FeedCardList from '../organisms/FeedCardList/FeedCardList';
 import Modal from '../organisms/Modal/Modal';
 import TopPanel from '../organisms/TopPanel/TopPanel';
 import useQuestionOwnerData from '../../hooks/useQuestionOwnerData';
+import useToggle from '../../hooks/useToggle';
 
 // Todo (송상훈) 좋아요 싫어요 로직구현, 무한스크롤 구현
 export default function QuestionPage() {
   const { imageSource, name, id } = useQuestionOwnerData();
   const { count, questions } = useQuestionData();
-
-  const [isModalClicked, setIsModalClicked] = useState(false);
-  const toggleModal = () => {
-    setIsModalClicked((prev) => !prev);
-  };
-
-  // 스크롤바 영역 보존
-  document.documentElement.style.scrollbarGutter = 'stable';
-
-  // 모달창이 띄워지면 스크롤 방지
-  useEffect(() => {
-    if (isModalClicked) {
-      document.body.style.overflow = 'hidden';
-    }
-    return () => {
-      document.body.style.overflow = 'unset';
-    };
-  }, [isModalClicked]);
+  const [isModalClicked, toggleModal] = useToggle();
 
   return (
     <>
